+++ src/loginServices.js
@@ -241,7 +241,7 @@ function loadResources(okapiUrl, store, tenant) {
     promises.push(discoverServices(store));
   }
 
-  Promise.all(promises).finally(() => store.dispatch(setOkapiReady()));
+  return Promise.all(promises);
 }
 
 /**
@@ -292,7 +292,7 @@ function createOkapiSession(okapiUrl, store, tenant, token, data) {
   localforage.setItem('okapiSess', okapiSess);
   store.dispatch(setSessionData(okapiSess));
 
-  loadResources(okapiUrl, store, tenant);
+  return loadResources(okapiUrl, store, tenant);
 }
 
 /**
@@ -309,24 +309,33 @@ function createOkapiSession(okapiUrl, store, tenant, token, data) {
  * @return Promise
  */
 function validateUser(okapiUrl, store, tenant, session) {
-  fetch(`${okapiUrl}/bl-users/_self`, { headers: getHeaders(tenant, session.token) }).then((resp) => {
+  return fetch(`${okapiUrl}/bl-users/_self`, { headers: getHeaders(tenant, session.token) }).then((resp) => {
     if (resp.status >= 400) {
       store.dispatch(clearCurrentUser());
       store.dispatch(clearOkapiToken());
-      store.dispatch(setOkapiReady());
       localforage.removeItem('okapiSess');
+      return null;
     } else {
       const { token, user, perms } = session;
       store.dispatch(setSessionData({ token, user, perms }));
-      loadResources(okapiUrl, store, tenant);
+      return loadResources(okapiUrl, store, tenant);
     }
   }).catch(() => {
     store.dispatch(setServerDown());
   });
 }
 
-function isSSOEnabled(okapiUrl, store, tenant) {
-  fetch(`${okapiUrl}/saml/check`, { headers: { 'X-Okapi-Tenant': tenant, 'Accept': 'application/json' } })
+/**
+ * getSSOEnabled
+ * return a promise that fetches from /saml/check and dispatches checkSSO.
+ * @param {*} okapiUrl
+ * @param {*} store
+ * @param {*} tenant
+ *
+ * @return Promise
+ */
+function getSSOEnabled(okapiUrl, store, tenant) {
+  return fetch(`${okapiUrl}/saml/check`, { headers: { 'X-Okapi-Tenant': tenant, 'Accept': 'application/json' } })
     .then((response) => {
       if (response.status >= 400) {
         store.dispatch(checkSSO(false));
@@ -335,11 +344,9 @@ function isSSOEnabled(okapiUrl, store, tenant) {
           store.dispatch(checkSSO(json.active));
         });
       }
-      store.dispatch(setOkapiReady());
     })
     .catch(() => {
       store.dispatch(checkSSO(false));
-      store.dispatch(setOkapiReady());
     });
 }
 
@@ -401,7 +408,11 @@ function processOkapiSession(okapiUrl, store, tenant, resp, ssoToken) {
   if (resp.status >= 400) {
     handleLoginError(dispatch, resp);
   } else {
-    resp.json().then(json => createOkapiSession(okapiUrl, store, tenant, token, json));
+    resp.json()
+      .then(json => createOkapiSession(okapiUrl, store, tenant, token, json))
+      .then(() => {
+        store.dispatch(setOkapiReady());
+      });
   }
   return resp;
 }
@@ -417,13 +428,16 @@ function processOkapiSession(okapiUrl, store, tenant, resp, ssoToken) {
  * @param {*} tenant
  */
 export function checkOkapiSession(okapiUrl, store, tenant) {
-  localforage.getItem('okapiSess').then((sess) => {
-    if (sess !== null) {
-      validateUser(okapiUrl, store, tenant, sess);
-    } else {
-      isSSOEnabled(okapiUrl, store, tenant);
-    }
-  });
+  localforage.getItem('okapiSess')
+    .then((sess) => {
+      return sess !== null ? validateUser(okapiUrl, store, tenant, sess) : null;
+    })
+    .then(() => {
+      getSSOEnabled(okapiUrl, store, tenant);
+    })
+    .finally(() => {
+      store.dispatch(setOkapiReady());
+    });
 }
 
 /**