+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 * Callouts are opened after page reload. Refs STCOR-518.
 * Use `==` where possible for more efficient queries. Refs STCOR-520.
 * Do not nest authentication forms. Refs STCOR-522.
+* When rehydrating a session from local storage, always dispatch `checkSSO`. Fixes STCOR-514.
 
 ## [6.0.0](https://github.com/folio-org/stripes-core/tree/v6.0.0) (2020-10-06)
 [Full Changelog](https://github.com/folio-org/stripes-core/compare/v5.0.2...v6.0.0)